+++ packages/react-ui/src/app/routes/platform/setup/copilot-providers/copilot-providers-config.ts
