+++ packages/react-ui/src/app/routes/platform/setup/copilot-providers/upsert-provider-dialog.tsx
